move test types, rename file, and add detail to comments
