+++ pkg/solana/codec/duration.go
