+++ pkg/solana/codec/solana_test.go
@@ -10,26 +10,28 @@ import (
 	"github.com/stretchr/testify/require"
 
 	codeccommon "github.com/smartcontractkit/chainlink-common/pkg/codec"
+	"github.com/smartcontractkit/chainlink-common/pkg/codec/encodings"
 	"github.com/smartcontractkit/chainlink-common/pkg/types"
 	"github.com/smartcontractkit/chainlink-common/pkg/utils/tests"
 
 	"github.com/smartcontractkit/chainlink-solana/pkg/solana/codec"
+	"github.com/smartcontractkit/chainlink-solana/pkg/solana/codec/testutils"
 )
 
 func TestNewIDLCodec(t *testing.T) {
 	t.Parallel()
 
 	ctx := tests.Context(t)
-	_, _, entry := codec.NewTestIDLAndCodec(t)
+	_, _, entry := newTestIDLAndCodec(t)
 
-	expected := codec.DefaultTestStruct
-	bts, err := entry.Encode(ctx, expected, codec.TestStructWithNestedStruct)
+	expected := testutils.DefaultTestStruct
+	bts, err := entry.Encode(ctx, expected, testutils.TestStructWithNestedStruct)
 
 	require.NoError(t, err)
 
-	var decoded codec.StructWithNestedStruct
+	var decoded testutils.StructWithNestedStruct
 
-	require.NoError(t, entry.Decode(ctx, bts, &decoded, codec.TestStructWithNestedStruct))
+	require.NoError(t, entry.Decode(ctx, bts, &decoded, testutils.TestStructWithNestedStruct))
 	require.Equal(t, expected, decoded)
 }
 
@@ -37,7 +39,7 @@ func TestNewIDLCodec_WithModifiers(t *testing.T) {
 	t.Parallel()
 
 	ctx := tests.Context(t)
-	_, _, idlCodec := codec.NewTestIDLAndCodec(t)
+	_, _, idlCodec := newTestIDLAndCodec(t)
 	modConfig := codeccommon.ModifiersConfig{
 		&codeccommon.RenameModifierConfig{Fields: map[string]string{"Value": "V"}},
 	}
@@ -45,15 +47,15 @@ func TestNewIDLCodec_WithModifiers(t *testing.T) {
 	renameMod, err := modConfig.ToModifier(codec.DecoderHooks...)
 	require.NoError(t, err)
 
-	idlCodecWithMods, err := codec.NewNamedModifierCodec(idlCodec, codec.TestStructWithNestedStruct, renameMod)
+	idlCodecWithMods, err := codec.NewNamedModifierCodec(idlCodec, testutils.TestStructWithNestedStruct, renameMod)
 	require.NoError(t, err)
 
 	type modifiedTestStruct struct {
 		V                uint8
-		InnerStruct      codec.ObjectRef1
+		InnerStruct      testutils.ObjectRef1
 		BasicNestedArray [][]uint32
 		Option           *string
-		DefinedArray     []codec.ObjectRef2
+		DefinedArray     []testutils.ObjectRef2
 		BasicVector      []string
 		TimeVal          int64
 		DurationVal      time.Duration
@@ -62,22 +64,22 @@ func TestNewIDLCodec_WithModifiers(t *testing.T) {
 	}
 
 	expected := modifiedTestStruct{
-		V:                codec.DefaultTestStruct.Value,
-		InnerStruct:      codec.DefaultTestStruct.InnerStruct,
-		BasicNestedArray: codec.DefaultTestStruct.BasicNestedArray,
-		Option:           codec.DefaultTestStruct.Option,
-		DefinedArray:     codec.DefaultTestStruct.DefinedArray,
-		BasicVector:      codec.DefaultTestStruct.BasicVector,
-		TimeVal:          codec.DefaultTestStruct.TimeVal,
-		DurationVal:      codec.DefaultTestStruct.DurationVal,
-		PublicKey:        codec.DefaultTestStruct.PublicKey,
-		EnumVal:          codec.DefaultTestStruct.EnumVal,
+		V:                testutils.DefaultTestStruct.Value,
+		InnerStruct:      testutils.DefaultTestStruct.InnerStruct,
+		BasicNestedArray: testutils.DefaultTestStruct.BasicNestedArray,
+		Option:           testutils.DefaultTestStruct.Option,
+		DefinedArray:     testutils.DefaultTestStruct.DefinedArray,
+		BasicVector:      testutils.DefaultTestStruct.BasicVector,
+		TimeVal:          testutils.DefaultTestStruct.TimeVal,
+		DurationVal:      testutils.DefaultTestStruct.DurationVal,
+		PublicKey:        testutils.DefaultTestStruct.PublicKey,
+		EnumVal:          testutils.DefaultTestStruct.EnumVal,
 	}
 
-	withModsBts, err := idlCodecWithMods.Encode(ctx, expected, codec.TestStructWithNestedStruct)
+	withModsBts, err := idlCodecWithMods.Encode(ctx, expected, testutils.TestStructWithNestedStruct)
 	require.NoError(t, err)
 
-	noModsBts, err := idlCodec.Encode(ctx, codec.DefaultTestStruct, codec.TestStructWithNestedStruct)
+	noModsBts, err := idlCodec.Encode(ctx, testutils.DefaultTestStruct, testutils.TestStructWithNestedStruct)
 
 	// the codec without modifiers should encode an unmodified struct to the same bytes
 	// as the codec with modifiers encodes a modified struct
@@ -87,14 +89,14 @@ func TestNewIDLCodec_WithModifiers(t *testing.T) {
 	var decoded modifiedTestStruct
 
 	// the codec with modifiers should decode from unmodified bytes into a modified struct
-	require.NoError(t, idlCodecWithMods.Decode(ctx, noModsBts, &decoded, codec.TestStructWithNestedStruct))
+	require.NoError(t, idlCodecWithMods.Decode(ctx, noModsBts, &decoded, testutils.TestStructWithNestedStruct))
 	require.Equal(t, expected, decoded)
 
-	var unmodifiedDecoded codec.StructWithNestedStruct
+	var unmodifiedDecoded testutils.StructWithNestedStruct
 
 	// the codec without modifiers should decode from unmodified bytes to the same values as
 	// modified struct
-	require.NoError(t, idlCodec.Decode(ctx, noModsBts, &unmodifiedDecoded, codec.TestStructWithNestedStruct))
+	require.NoError(t, idlCodec.Decode(ctx, noModsBts, &unmodifiedDecoded, testutils.TestStructWithNestedStruct))
 	require.Equal(t, expected.V, unmodifiedDecoded.Value)
 	require.Equal(t, expected.TimeVal, unmodifiedDecoded.TimeVal)
 	require.Equal(t, expected.DurationVal, unmodifiedDecoded.DurationVal)
@@ -105,51 +107,8 @@ func TestNewIDLCodec_WithModifiers(t *testing.T) {
 func TestNewIDLCodec_CircularDependency(t *testing.T) {
 	t.Parallel()
 
-	rawIDL := `{
-		"accounts": [{
-			"name": "TopLevelStruct",
-			"type": {
-				"kind": "struct",
-				"fields": [{
-					"name": "circularOne",
-					"type": {
-						"defined": "TypeOne"
-					}
-				}, {
-					"name": "circularTwo",
-					"type": {
-						"defined": "TypeTwo"
-					}
-				}]
-			}
-		}],
-		"types": [{
-			"name": "TypeOne",
-			"type": {
-				"kind": "struct",
-				"fields": [{
-					"name": "circular",
-					"type": {
-						"defined": "TypeTwo"
-					}
-				}]
-			}
-		}, {
-			"name": "TypeTwo",
-			"type": {
-				"kind": "struct",
-				"fields": [{
-					"name": "circular",
-					"type": {
-						"defined": "TypeOne"
-					}
-				}]
-			}
-		}]
-	}`
-
 	var idl codec.IDL
-	if err := json.Unmarshal([]byte(rawIDL), &idl); err != nil {
+	if err := json.Unmarshal([]byte(testutils.CircularDepIDL), &idl); err != nil {
 		t.Logf("failed to unmarshal test IDL: %s", err.Error())
 		t.FailNow()
 	}
@@ -157,4 +116,24 @@ func TestNewIDLCodec_CircularDependency(t *testing.T) {
 	_, err := codec.NewIDLCodec(idl)
 
 	assert.ErrorIs(t, err, types.ErrInvalidConfig)
+}
+
+func newTestIDLAndCodec(t *testing.T) (string, codec.IDL, encodings.CodecFromTypeCodec) {
+	t.Helper()
+
+	var idl codec.IDL
+	if err := json.Unmarshal([]byte(testutils.JSONIDLWithAllTypes), &idl); err != nil {
+		t.Logf("failed to unmarshal test IDL: %s", err.Error())
+		t.FailNow()
+	}
+
+	entry, err := codec.NewIDLCodec(idl)
+	if err != nil {
+		t.Logf("failed to create new codec from test IDL: %s", err.Error())
+		t.FailNow()
+	}
+
+	require.NotNil(t, entry)
+
+	return testutils.JSONIDLWithAllTypes, idl, entry
 }
