+++ pkg/solana/codec/testutils/testIDL.json
@@ -1,0 +1,159 @@
+{
+  "version": "0.1.0",
+  "name": "some_test_idl",
+  "accounts": [
+    {
+      "name": "StructWithNestedStruct",
+      "type": {
+        "kind": "struct",
+        "fields": [
+          {
+            "name": "value",
+            "type": "u8"
+          },
+          {
+            "name": "innerStruct",
+            "type": {
+              "defined": "ObjectRef1"
+            }
+          },
+          {
+            "name": "basicNestedArray",
+            "type": {
+              "array": [
+                {
+                  "array": [
+                    "u32",
+                    3
+                  ]
+                },
+                3
+              ]
+            }
+          },
+          {
+            "name": "option",
+            "type": {
+              "option": "string"
+            }
+          },
+          {
+            "name": "definedArray",
+            "type": {
+              "array": [
+                {
+                  "defined": "ObjectRef2"
+                },
+                2
+              ]
+            }
+          },
+          {
+            "name": "basicVector",
+            "type": {
+              "vec": "string"
+            }
+          },
+          {
+            "name": "timeVal",
+            "type": "unixTimestamp"
+          },
+          {
+            "name": "durationVal",
+            "type": "duration"
+          },
+          {
+            "name": "publicKey",
+            "type": "publicKey"
+          },
+          {
+            "name": "enumVal",
+            "type": {
+              "defined": "SimpleEnum"
+            }
+          }
+        ]
+      }
+    }
+  ],
+  "types": [
+    {
+      "name": "ObjectRef1",
+      "type": {
+        "kind": "struct",
+        "fields": [
+          {
+            "name": "prop1",
+            "type": "i8"
+          },
+          {
+            "name": "prop2",
+            "type": "string"
+          },
+          {
+            "name": "prop3",
+            "type": "u128"
+          },
+          {
+            "name": "prop4",
+            "type": "u16"
+          },
+          {
+            "name": "prop5",
+            "type": "u64"
+          },
+          {
+            "name": "prop6",
+            "type": "bool"
+          }
+        ]
+      }
+    },
+    {
+      "name": "ObjectRef2",
+      "type": {
+        "kind": "struct",
+        "fields": [
+          {
+            "name": "prop1",
+            "type": "u32"
+          },
+          {
+            "name": "prop2",
+            "type": "i128"
+          },
+          {
+            "name": "prop3",
+            "type": "i16"
+          },
+          {
+            "name": "prop4",
+            "type": "i32"
+          },
+          {
+            "name": "prop5",
+            "type": "i64"
+          },
+          {
+            "name": "prop6",
+            "type": "bytes"
+          }
+        ]
+      }
+    },
+    {
+      "name": "SimpleEnum",
+      "type": {
+        "kind": "enum",
+        "variants": [
+          {
+            "name": "A"
+          },
+          {
+            "name": "B"
+          }
+        ]
+      }
+    }
+  ]
+}
