+++ pkg/solana/config/chain_reader.go
@@ -29,7 +29,9 @@ type chainDataProcedureFields struct {
 	IDLAccount string `json:"idlAccount"`
 	// Type describes the procedure type to use such as internal for static values,
 	// anchor-read for using an anchor generated IDL to read values from an account,
-	// or custom structure for reading from a native account.
+	// or custom structure for reading from a native account. Currently, only anchor
+	// reads are supported, but the type is a placeholder to allow internal functions
+	// to be run apart from anchor reads.
 	Type ProcedureType `json:"type"`
 	// OutputModifications provides modifiers to convert chain data format to custom
 	// output formats.
